fix(*): refactor, simplify, remove lots of deps
