+++ .helarc.yml
@@ -1,1 +1,1 @@
-extends: ./src/index-cjs.js
+extends: ./src/index.js
