+++ .npmrc
@@ -1,4 +1,5 @@
-registry=http://registry.npmjs.org/
 save-exact=true
 save=true
 username=olstenlarck
+silent=true
+progress=true
