+++ .prettierrc.json
@@ -1,7 +1,7 @@
 {
   "tabWidth": 2,
-  "printWidth": 80,
-  "semi": false,
+  "printWidth": 90,
+  "semi": true,
   "useTabs": false,
   "singleQuote": true,
   "bracketSpacing": true,
