+++ .travis.yml
@@ -18,8 +18,6 @@ before_install:
   - curl -o- -L https://yarnpkg.com/install.sh | bash
   - export PATH=$HOME/.yarn/bin:$PATH
 
-before_script: yarn link
-
 script: yarn test
 
 after_success:
