+++ .verb.md
@@ -64,7 +64,7 @@ See available bundles at [`https://unpkg.com/{%= name %}/dist/browser/`](https:/
 ## API
 Review carefully the provided examples and the working [tests](./test).
 
-{%= apidocs('src/index.js') %}
+{%= apidocs('src/index.esm.js') %}
 
 **[back to top](#thetop)**
