+++ USERS.md
@@ -3,4 +3,4 @@
 If you or your company uses this project, please feel free to add your name to
 this list as bullet point! Eventually we may have a website to showcase these (wanna build it!?)
 
-- [tunnckoCore](https://github.com/tunnckoCore) uses it to scaffold and manage all of its other projects.
+- add bullets here...
